requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.5
pydantic>=2.5.2
pandas>=2.0.0
openpyxl>=3.0.0
langchain[azure]>=0.1.0
pypdf2>=3.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
tabula-py>=2.7.0
scikit-learn>=1.0.2
nltk>=3.8.1
azure-identity>=1.12.0
azure-ai-textanalytics>=5.3.0
